Add files via upload
This code is regarding an ATM machine .It's user interface is designed with options such as withdrawing, depositing and checking balance.  Each option  is implemented by methods such as withdraw(amount), deposit(amount), and checkbalance(). Appropriate messages to the users are displayed based on their chosen options and the success or failure of their transactions are also displayed.